Nodes Recovery: recover nodes saved in DB when restarting RM
- implement node and infrastructure recovery: when node to recover can be looked up,
they are re-added to the RM and to the node source, and when they cannot be looked up,
they are recreated and notified as down
- create a onDownNode hook in the infrastructure manager that is called each time a node is detected as down.
Move logic that handles down nodes from the remove method to the onDownNode method for the LocalInfrastructure and the HostsFileBasedInfrastructure.
- remove a test in NodeSourceTest that was not relevant anymore
- moved tests from NodesLockRestorationManagerTest to NodesRecoveryManagerTest
- add tests for LocalInfrastructure where either nodes can be looked up or not
- add test for SSHInfrastructureV2 where nodes that are detected down when recovering
- change behavior of LocalInfrastructure and SSHInfrastructureV2 when node are not present when recovering: down nodes are now redeployed upon restart
- add a property to indicate if nodes should be removed and shutdown if RM is shutdown. Make the process tree killer behave accordingly- make termination of RM os-independent in tests
- add test covering the case where nodes of an SSHInfrastructureV2 can be looked up after a recovery. This requires changing the setup of the ssh server that is instanciated for the test.

+++ rm/rm-server/src/test/java/functionaltests/utils/TestRM.java
@@ -34,7 +34,6 @@ import java.util.List;
 
 import org.objectweb.proactive.core.config.CentralPAPropertyRepository;
 import org.objectweb.proactive.extensions.pnp.PNPConfig;
-import org.objectweb.proactive.utils.OperatingSystem;
 import org.ow2.proactive.resourcemanager.authentication.RMAuthentication;
 import org.ow2.proactive.resourcemanager.core.properties.PAResourceManagerProperties;
 import org.ow2.proactive.resourcemanager.frontend.RMConnection;
@@ -74,7 +73,8 @@ public class TestRM {
                                configurationFile == null && startedConfiguration.equals(DEFAULT_CONFIGURATION));
     }
 
-    public synchronized void start(String configurationFile, int pnpPort, String... jvmArgs) throws Exception {
+    public synchronized void start(String configurationFile, int pnpPort, String classpath, String... jvmArgs)
+            throws Exception {
         if (configurationFile == null) {
             configurationFile = DEFAULT_CONFIGURATION;
         }
@@ -118,7 +118,7 @@ public class TestRM {
         commandLine.add(CentralPAPropertyRepository.PA_RUNTIME_PING.getCmdLine() + false);
 
         commandLine.add("-cp");
-        commandLine.add(testClasspath());
+        commandLine.add(classpath);
         commandLine.add("-Djava.library.path=" + System.getProperty("java.library.path"));
         commandLine.add(CentralPAPropertyRepository.PA_TEST.getCmdLine() + "true");
         commandLine.add("-Djava.awt.headless=true"); // For Mac builds
@@ -129,8 +129,12 @@ public class TestRM {
 
         ProcessBuilder processBuilder = new ProcessBuilder(commandLine);
         processBuilder.redirectErrorStream(true);
-        processTreeKiller = CookieBasedProcessTreeKiller.createProcessChildrenKiller("TEST_RM",
-                                                                                     processBuilder.environment());
+
+        if (!PAResourceManagerProperties.RM_PRESERVE_NODES_ON_EXIT.getValueAsBoolean()) {
+            processTreeKiller = CookieBasedProcessTreeKiller.createProcessChildrenKiller("TEST_RM",
+                                                                                         processBuilder.environment());
+        }
+
         rmProcess = processBuilder.start();
 
         InputStreamReaderThread outputReader = new InputStreamReaderThread(rmProcess.getInputStream(), "[RM output]: ");
@@ -145,24 +149,15 @@ public class TestRM {
             System.out.println("Destroying RM process.");
             rmProcess.destroy();
             rmProcess.waitFor();
-            processTreeKiller.kill();
+            if (!PAResourceManagerProperties.RM_PRESERVE_NODES_ON_EXIT.getValueAsBoolean()) {
+                processTreeKiller.kill();
+            }
             rmProcess = null;
         }
     }
 
     public String getUrl() {
         return "pnp://localhost:" + pnpPort + "/";
-    }
-
-    private static String testClasspath() {
-        String home = PAResourceManagerProperties.RM_HOME.getValueAsString();
-        String classpathToLibFolderWithWildcard = home + File.separator + "dist" + File.separator + "lib" +
-                                                  File.separator + "*";
-        if (OperatingSystem.getOperatingSystem().equals(OperatingSystem.windows)) {
-            // required by windows otherwise wildcard is expanded
-            classpathToLibFolderWithWildcard = "\"" + classpathToLibFolderWithWildcard + "\"";
-        }
-        return classpathToLibFolderWithWildcard;
     }
 
     public synchronized RMAuthentication getAuth() {
